Make Screen::showCursor() accept focused widget
Accepts only Window widgets.  This fixes an issue when cursor is moved
as a result of drawing other widgets after the current one.  If Prompt
is the bottom, the bug won't be noticeable as it will be drawn last and
place cursor where it should be right before screen redrawing.